@@ -2,8 +2,8 @@
   "name": "@skysmack/ng-framework",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
     "@skysmack/framework": "*",
     "@skysmack/redux": "*",
     "parse-link-header": "^1.0.1"
