@@ -2,8 +2,8 @@
   "name": "@skysmack/ng-packages",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
     "@skysmack/ng-framework": "*",
     "@skysmack/ng-ui": "*",
     "@skysmack/framework": "*",
