@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { StandardSettingsActions } from './standard-settings-actions';
 import { Settings } from '@skysmack/ng-ui';
-import { map } from 'rxjs/operators';
+import { map, filter } from 'rxjs/operators';
 
 @Injectable({ providedIn: 'root' })
 export class StandardSettingsRedux {
@@ -25,7 +25,11 @@ export class StandardSettingsRedux {
     public getSettings = (): Observable<Settings> => {
         return this.ngRedux.select((state: any) => state).pipe(
             map(state => {
-                return state.standardSettings.settings;
+                const temp = {
+                    language: 'en',
+                    tenantUrl: ''
+                };
+                return state.standardSettings ? state.standardSettings.settings : temp;
             })
         );
     }
