@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { AccountModule } from '@skysmack/portal-packages';
+import { AccountModule } from './../../../../../../../libs/portal-packages/src/lib/account/account.module';
 
 @NgModule({
   imports: [
