@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
-import { InvoicesCashPaymentsModule } from '@skysmack/portal-packages';
-
+import { InvoicesCashPaymentsModule } from './../../../../../../../libs/portal-packages/src/lib/invoices-cash-payments/invoices-cash-payments.module';
 
 @NgModule({
   imports: [
