@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
-import { InvoicesModule } from '@skysmack/portal-packages';
-
+import { InvoicesModule } from './../../../../../../../libs/portal-packages/src/lib/invoices/invoices.module';
 
 @NgModule({
   imports: [
