@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { LodgingReservationsModule } from '@skysmack/portal-packages';
+import { LodgingReservationsModule } from './../../../../../../../libs/portal-packages/src/lib/lodging-reservations/lodging-reservations.module';
 
 @NgModule({
   imports: [
