@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
-import { LodgingsModule } from '@skysmack/portal-packages';
+import { LodgingsModule } from './../../../../../../../libs/portal-packages/src/lib/lodgings/lodgings.module';
 
 @NgModule({
   imports: [
     LodgingsModule
   ]
 })
-export class LodgingsWrapperModule {}
+export class LodgingsWrapperModule { }
