@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
-import { Oauth2Module } from '@skysmack/portal-packages';
+import { Oauth2Module } from './../../../../../../../libs/portal-packages/src/lib/oauth2/oauth2.module';
 
 @NgModule({
   imports: [
     Oauth2Module
   ]
 })
-export class OAuth2WrapperModule {}
+export class OAuth2WrapperModule { }
