@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { ProductsPricingsModule } from '@skysmack/portal-packages';
+import { ProductsPricingsModule } from './../../../../../../../libs/portal-packages/src/lib/products-pricings/products-pricings.module';
 
 @NgModule({
   imports: [
