@@ -1,10 +1,9 @@
 import { NgModule } from '@angular/core';
-import { TerminalPaymentsModule } from '@skysmack/portal-packages';
-
+import { TerminalPaymentsModule } from './../../../../../../../libs/portal-packages/src/lib/terminal-payments/terminal-payments.module';
 
 @NgModule({
   imports: [
     TerminalPaymentsModule
   ]
 })
-export class TerminalPaymentsWrapperModule {}
+export class TerminalPaymentsWrapperModule { }
