@@ -4,7 +4,7 @@ import { PackagesType } from '@skysmack/ng-core';
 export class PackagesPackageManifest extends PackagesType implements PackageManifest {
     public icon = 'apps';
     public menuLocation = 'docker';
-    public modulePath = './../packages/modules/packages_wrapper.module#PackagesWrapperModule';
+    public modulePath = './../../../../../../libs/portal-packages/src/lib/packages#PackagesModule';
 }
 
 export function loadPackagesPackage(packageLoader: PackageLoader) {
