@@ -2,6 +2,9 @@ import { Routes } from '@angular/router';
 import { ChangePasswordComponent } from './change-password/change-password.component';
 import { AccountComponent } from './account/account.component';
 import { ConfirmEmailComponent } from './confirm-email/confirm-email.component';
+import { ForgotPasswordComponent } from './forgot-password/forgot-password.component';
+import { RecoverPasswordComponent } from './recover-password/recover-password.component';
+
 
 export const accountsRoutes: Routes = [
     {
@@ -11,6 +14,14 @@ export const accountsRoutes: Routes = [
             },
             {
                 path: 'confirm-email', component: ConfirmEmailComponent
+            },
+            {
+                path: 'forgot-password',
+                component: ForgotPasswordComponent
+            },
+            {
+                path: 'recover-password',
+                component: RecoverPasswordComponent
             }
         ]
     }
@@ -19,5 +30,7 @@ export const accountsRoutes: Routes = [
 export const accountsComponents: any[] = [
     AccountComponent,
     ChangePasswordComponent,
-    ConfirmEmailComponent
+    ConfirmEmailComponent,
+    ForgotPasswordComponent,
+    RecoverPasswordComponent
 ];
