@@ -1,2 +1,6 @@
 export * from './accounts-components';
-
+export * from './account/account.component';
+export * from './change-password/change-password.component';
+export * from './confirm-email/confirm-email.component';
+export * from './forgot-password/forgot-password.component';
+export * from './recover-password/recover-password.component';
