@@ -1,24 +1,37 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { NgRecoverPasswordFieldsConfig } from './ng-recover-password-fields-config';
 import { FormHelper } from '@skysmack/ng-ui';
-import { Router } from '@angular/router';
+import { Router, ActivatedRoute } from '@angular/router';
 import { toLocalObject } from '@skysmack/framework';
+import { BaseComponent, EditorNavService } from '@skysmack/portal-ui';
+import { AccountAppState } from '@skysmack/packages-account';
+import { NgSkysmackStore } from '@skysmack/ng-core';
+import { NgAccountRequests } from '@skysmack/ng-packages';
 
 @Component({
   selector: 'ss-recover-password',
   templateUrl: './recover-password.component.html',
   styleUrls: ['./recover-password.component.scss']
 })
-export class RecoverPasswordComponent implements OnInit, OnDestroy {
+export class RecoverPasswordComponent extends BaseComponent<AccountAppState, unknown> implements OnInit, OnDestroy {
 
   public fields$;
 
   constructor(
+    public router: Router,
+    public activatedRoute: ActivatedRoute,
+    public skysmackStore: NgSkysmackStore,
     public fieldsConfig: NgRecoverPasswordFieldsConfig,
-    public router: Router
-  ) { }
+    public editorNavService: EditorNavService,
+    public accountRequest: NgAccountRequests
+  ) {
+    super(router, activatedRoute, skysmackStore);
+  }
 
   ngOnInit() {
+    super.ngOnInit();
+    this.editorNavService.showEditorNav();
+
     const token = this.router.url.split('=')[1];
     if (token) {
       const tokenObject = toLocalObject(token);
@@ -26,11 +39,11 @@ export class RecoverPasswordComponent implements OnInit, OnDestroy {
     } else {
       this.fields$ = this.fieldsConfig.getFields(undefined);
     }
-
   }
 
   ngOnDestroy() {
-
+    super.ngOnDestroy();
+    this.editorNavService.hideEditorNav();
   }
 
   public onSubmit(fh: FormHelper) {
