@@ -1,4 +1,4 @@
-import { NativeDateAdapter } from '@angular/material';
+import { NativeDateAdapter } from '@angular/material/core';
 import { Injectable } from '@angular/core';
 import * as _moment from 'moment';
 const moment = _moment;
