@@ -1,5 +1,5 @@
-export * from './recurring-custom-expression';
-export * from './recurring-daily';
-export * from './recurring-monthly';
-export * from './recurring-weekly';
-export * from './recurring-yearly';
+export * from './recurring-custom-expression/index';
+export * from './recurring-daily/index';
+export * from './recurring-monthly/index';
+export * from './recurring-weekly/index';
+export * from './recurring-yearly/index';
