@@ -1,4 +1,5 @@
-export * from './models';
+export * from './models/index';
 export * from './components/index';
+
 export * from './recurring-expression-field.component';
 export * from './recurring-expression-field.module';
