@@ -1,31 +1,29 @@
 import { NgModule } from '@angular/core';
 import { MatMomentDateModule } from '@angular/material-moment-adapter';
-import {
-  MatGridListModule,
-  MatProgressSpinnerModule,
-  MatCardModule,
-  MatToolbarModule,
-  MatButtonModule,
-  MatMenuModule,
-  MatIconModule,
-  MatDatepickerModule,
-  MatFormFieldModule,
-  MatInputModule,
-  MatSidenavModule,
-  MatCheckboxModule,
-  MatTooltipModule,
-  MatSelectModule,
-  MatSlideToggleModule,
-  MatButtonToggleModule,
-  MatTabsModule,
-  MatTableModule,
-  MatSortModule,
-  MatPaginatorModule,
-  MatProgressBarModule,
-  MatSnackBarModule,
-  MatBadgeModule,
-  MatDividerModule
-} from '@angular/material';
+import { MatBadgeModule } from '@angular/material/badge';
+import { MatButtonModule } from '@angular/material/button';
+import { MatButtonToggleModule } from '@angular/material/button-toggle';
+import { MatCardModule } from '@angular/material/card';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatDividerModule } from '@angular/material/divider';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatGridListModule } from '@angular/material/grid-list';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatSelectModule } from '@angular/material/select';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatSortModule } from '@angular/material/sort';
+import { MatTableModule } from '@angular/material/table';
+import { MatTabsModule } from '@angular/material/tabs';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatTooltipModule } from '@angular/material/tooltip';
 import { CdkAccordionModule } from '@angular/cdk/accordion';
 import { ScrollDispatchModule } from '@angular/cdk/scrolling';
 
